chore: ini query src
